Change Encode_UTF8 API
The "len" argument of Encode_UTF8() was used as the source length on input and
as the result length on output. This caused some confusion in the calling code
about the meaning of a variable before and after the call. Now "len" is updated
to the amount of source used, and the result length is in the function return
value. The meaning of "len" argument and return value of RL_Encode_UTF8() is
changed accordingly. This change fixes bugs involving calls to Encode_UTF8().

Also, "uni" and "ccr" arguments of Encode_UTF8() and Length_As_UTF8() are
replaced with one "opts" argument. REBUNI* source is signalled by
"ENC_OPT_UNISRC" flag, and CRLF handling is signalled by "ENC_OPT_CRLF" flag.
Make_UTF8_Binary() helper function is added that calls Length_As_UTF8(),
Make_Binary(), Encode_UTF8().